refac: Remove remaining file based calculation results
