+++ source/databricks/calculation-engine/contracts/calculation-file-paths.yml
@@ -3,27 +3,6 @@
 # The contract uses regular expressions to specify path and filename patterns.
 # Due to the distributed nature of Spark the file names cannot be determined precisely.
 
-result_file:
-    # Example: calculation-output/batch_id=0b15a420-9fc8-409a-a169-fbd49479d718/result/grouping=es_ga/time_series_type=production/grid_area=806/gln=1234567/
-    # Example: calculation-output/batch_id=0b15a420-9fc8-409a-a169-fbd49479d718/result/grouping=brp_ga/time_series_type=production/grid_area=806/gln=1234567/
-    directory_expression: calculation-output/batch_id=\w{8}-\w{4}-\w{4}-\w{4}-\w{12}/result/grouping=.*/time_series_type=.*/grid_area=\d{3}/gln=(\d{13}|\d{16})/
-    extension: .json
-
-result_file_for_total_grid_area:
-    # Example: calculation-output/batch_id=0b15a420-9fc8-409a-a169-fbd49479d718/result/total_ga/time_series_type=production/grid_area=806/
-    directory_expression: calculation-output/batch_id=\w{8}-\w{4}-\w{4}-\w{4}-\w{12}/result/grouping=.*/time_series_type=.*/grid_area=\d{3}/
-    extension: .json
-
-result_file_for_ga_brp_es:
-    # Example: calculation-output/batch_id=0b15a420-9fc8-409a-a169-fbd49479d718/result/es_brp_ga/time_series_type=production/grid_area=806/balance_responsible_party_gln=1234567/energy_supplier_gln=1234567//
-    directory_expression: calculation-output/batch_id=\w{8}-\w{4}-\w{4}-\w{4}-\w{12}/result/grouping=es_brp_ga/time_series_type=.*/grid_area=\d{3}/balance_responsible_party_gln=(\d{13}|\d{16})/energy_supplier_gln=(\d{13}|\d{16})/
-    extension: .json
-
-actors_file:
-    # Example: calculation-output/batch_id=0b15a420-9fc8-409a-a169-fbd49479d718/actors/time_series_type=non_profiled_consumption/grid_area=806/
-    directory_expression: calculation-output/batch_id=\w{8}-\w{4}-\w{4}-\w{4}-\w{12}/actors/time_series_type=.*/grid_area=\d{3}/
-    extension: .json
-
 time_series_hour_basis_data_file_for_total_grid_area:
     # Example: calculation-output/batch_id=0b15a420-9fc8-409a-a169-fbd49479d718/basis_data/time-series-hour/grouping=total_ga/grid_area=806/
     directory_expression: calculation-output/batch_id=\w{8}-\w{4}-\w{4}-\w{4}-\w{12}/basis_data/time_series_hour/grouping=total_ga/grid_area=\d{3}/
