--- source/databricks/calculation-engine/contracts/calculator-actor.json
@@ -1,13 +1,0 @@
-{
-    "fields": [
-      {
-        "name": "energy_supplier_gln",
-        "type": "string"
-      },
-      {
-        "name": "balance_responsible_party_gln",
-        "type": "string"
-      }
-    ]
-  }
-
