--- source/databricks/calculation-engine/contracts/calculator-file-result.json
@@ -1,17 +1,0 @@
-{
-    "fields": [
-      {
-        "name": "quantity",
-        "type": "string"
-      },
-      {
-        "name": "quality",
-        "type": "string"
-      },
-      {
-        "name": "quarter_time",
-        "type": "string"
-      }
-    ]
-  }
-
