--- source/databricks/calculation-engine/package/shared/period.py
@@ -1,32 +1,0 @@
-# Copyright 2020 Energinet DataHub A/S
-#
-# Licensed under the Apache License, Version 2.0 (the "License2");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from datetime import datetime
-from argparse import Namespace
-import dateutil.parser
-
-
-class Period:
-    from_date: datetime
-    to_date: datetime
-
-    def __init__(self, from_date: datetime, to_date: datetime):
-        self.from_date = from_date
-        self.to_date = to_date
-
-
-def parse_period(beginning_date_time: str, end_date_time: str) -> Period:
-    period = Period(
-        dateutil.parser.parse(beginning_date_time), dateutil.parser.parse(end_date_time)
-    )
-    return period
