--- source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/Interfaces/Actors/actor-list.json
@@ -1,2 +1,0 @@
-{"gln":"5790001102890"}
-{"gln":"5790000013296"}
