+++ source/dotnet/wholesale-api/WebApi.IntegrationTests/Fixtures/WebApi/WebApiFactory.cs
@@ -14,7 +14,6 @@
 
 using Energinet.DataHub.Core.App.Common.Abstractions.Users;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
-using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.Actors;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.SettlementReports;
 using Energinet.DataHub.Wholesale.Common.Security;
 using Microsoft.AspNetCore.Authorization;
@@ -59,8 +58,6 @@ public class WebApiFactory : WebApplicationFactory<Startup>
             services.AddScoped(_ =>
                 SettlementReportApplicationServiceMock?.Object ??
                 new Mock<ISettlementReportClient>().Object);
-            services.AddScoped(_ =>
-                ActorRepositoryMock?.Object ?? new Mock<IActorClient>().Object);
             services.AddScoped(_ => BatchesClientMock?.Object ?? new Mock<IBatchesClient>().Object);
 
             var defaultUserContext = new Mock<IUserContext<FrontendUser>>();
@@ -74,8 +71,6 @@ public class WebApiFactory : WebApplicationFactory<Startup>
     /// NOTE: This will only work as expected as long as no tests are executed in parallel.
     /// </summary>
     public Mock<ISettlementReportClient>? SettlementReportApplicationServiceMock { get; set; }
-
-    public Mock<IActorClient>? ActorRepositoryMock { get; set; }
 
     public Mock<IBatchesClient>? BatchesClientMock { get; set; }
