+++ source/dotnet/wholesale-api/WebApi.IntegrationTests/WebApi/V3/BalanceResponsiblePartyCalculationResultTests.cs
@@ -13,16 +13,11 @@
 // limitations under the License.
 
 using System.Net;
-using System.Net.Http.Json;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
-using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.Actors;
-using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.Actors.Model;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.TestCommon.Fixture.WebApi;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.WebApi;
-using Energinet.DataHub.Wholesale.WebApi.V3;
 using FluentAssertions;
-using Moq;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -54,30 +49,5 @@ public class BalanceResponsiblePartyCalculationResultTests : WebApiTestBase
 
         // Assert
         actualContent.StatusCode.Should().Be(expectedHttpStatusCode);
-    }
-
-    [Theory]
-    [InlineAutoMoqData]
-    public async Task HTTP_GET_V3_ReturnsExpectedActorInJson(
-        Mock<IActorClient> applicationServiceMock,
-        Guid batchId,
-        string gridAreaCode,
-        TimeSeriesType timeSeriesType,
-        Actor expectedActor)
-    {
-        // Arrange
-        var url = $"/v3/batches/{batchId}/processes/{gridAreaCode}/time-series-types/{timeSeriesType}/balance-responsible-parties";
-
-        applicationServiceMock
-            .Setup(service => service.GetBalanceResponsiblePartiesAsync(batchId, gridAreaCode, timeSeriesType))
-            .ReturnsAsync(() => new[] { expectedActor });
-        Factory.ActorRepositoryMock = applicationServiceMock;
-
-        // Act
-        var actualContent = await Client.GetAsync(url);
-
-        // Assert
-        var actualActors = await actualContent.Content.ReadFromJsonAsync<List<ActorDto>>();
-        actualActors!.Single().Should().BeEquivalentTo(expectedActor);
     }
 }
