+++ source/dotnet/wholesale-api/WebApi.UnitTests/WebApi.UnitTests.csproj
@@ -63,9 +63,6 @@ limitations under the License.
     <EmbeddedResource Include="..\..\..\databricks\calculation-engine\contracts\calculation-job-parameters-reference.txt">
       <Link>Infrastructure\JobRunner\calculation-job-parameters-reference.txt</Link>
     </EmbeddedResource>
-    <EmbeddedResource Include="..\..\..\databricks\calculation-engine\contracts\calculator-file-result.json">
-      <Link>Infrastructure\Integration\CalculationOutput\calculator-file-result.json</Link>
-    </EmbeddedResource>
     <EmbeddedResource Include="..\..\..\databricks\calculation-engine\contracts\enums\quantity-quality.json">
       <Link>Infrastructure\Integration\CalculationOutput\time-series-point-quality.json</Link>
     </EmbeddedResource>
@@ -74,12 +71,6 @@ limitations under the License.
   <ItemGroup>
     <EmbeddedResource Include="..\..\..\databricks\calculation-engine\contracts\calculation-file-paths.yml">
       <Link>Infrastructure\BasisData\calculation-file-paths.yml</Link>
-    </EmbeddedResource>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="..\..\..\databricks\calculation-engine\contracts\calculator-actor.json">
-      <Link>Infrastructure\Actor\calculator-actor.json</Link>
     </EmbeddedResource>
   </ItemGroup>
