--- source/dotnet/wholesale-api/WebApi/V3/BalanceResponsiblePartyCalculationResult/BalanceResponsiblePartyCalculationResultController.cs
@@ -1,50 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.Actors;
-using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Energinet.DataHub.Wholesale.WebApi.V3.BalanceResponsiblePartyCalculationResult;
-
-/// <summary>
-/// Energy suppliers for which batch results have been calculated.
-/// </summary>
-[Route("/v3/batches/{batchId}/processes/{gridAreaCode}/time-series-types/{timeSeriesType}/balance-responsible-parties")]
-public class BalanceResponsiblePartyCalculationResultController : V3ControllerBase
-{
-    private readonly IActorClient _actorClient;
-
-    public BalanceResponsiblePartyCalculationResultController(IActorClient actorClient)
-    {
-        _actorClient = actorClient;
-    }
-
-    /// <summary>
-    /// Balance responsible parties.
-    /// </summary>
-    [HttpGet(Name = "GetListOfBalanceResponsibleParties")]
-    [Produces("application/json", Type = typeof(List<ActorDto>))]
-    public async Task<List<ActorDto>> GetAllAsync(
-        [FromRoute] Guid batchId,
-        [FromRoute] string gridAreaCode,
-        [FromRoute] TimeSeriesType timeSeriesType)
-    {
-        var balanceResponsibleParties = await _actorClient.GetBalanceResponsiblePartiesAsync(batchId, gridAreaCode, timeSeriesType).ConfigureAwait(false);
-
-        return balanceResponsibleParties
-            .Select(a => new ActorDto(a.Gln))
-            .ToList();
-    }
-}
